# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Star Wars REST API with Swagger",
    "version": "0.1.0",
    "description": "Welcome to the swapi, the Star Wars API! This documentation should help you familiarise yourself with the resources available and how to consume them with HTTP requests.",
    "license": {
      "name": "MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "contact": {
      "name": "Corentin Goujon-Cholet",
      "email": "corentingcholet@gmail.com"
    }
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Planet": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "diameter": {
            "type": "string"
          },
          "rotation_period": {
            "type": "string"
          },
          "orbital_period": {
            "type": "string"
          },
          "gravity": {
            "type": "string"
          },
          "population": {
            "type": "string"
          },
          "climate": {
            "type": "string"
          },
          "terrain": {
            "type": "string"
          },
          "surface_water": {
            "type": "string"
          },
          "residents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "film": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "url": {
            "type": "string"
          }
        }
      },
      "Specie": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "classification": {
            "type": "string"
          },
          "designation": {
            "type": "string"
          },
          "average_height": {
            "type": "string"
          },
          "average_lifespan": {
            "type": "string"
          },
          "eye_colors": {
            "type": "string"
          },
          "hair_colors": {
            "type": "string"
          },
          "skin_colors": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "homeworld": {
            "type": "string"
          },
          "people": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "films": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "url": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
...
    "/api/planet": {
      "post": {
        "tags": [
          "Planet"
        ],
        "description": "Get all planets",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Planet"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Planet"
        ],
        "description": "Get all planet",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Planet"
        ],
        "description": "Delete all Planets",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/api/planet/{id}": {
      "get": {
        "tags": [
          "Planet"
        ],
        "description": "Get planet by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The planet id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "put": {
        "tags": [
          "Planet"
        ],
        "description": "Modify planet by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The planet id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Planet"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Planet"
        ],
        "description": "Delete planet by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The planet id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/api/specie": {
      "post": {
        "tags": [
          "Specie"
        ],
        "description": "Get all species",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Specie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Specie"
        ],
        "description": "Get all specie",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Specie"
        ],
        "description": "Delete all Species",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/api/specie/{id}": {
      "get": {
        "tags": [
          "Specie"
        ],
        "description": "Get specie by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The specie id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "put": {
        "tags": [
          "Specie"
        ],
        "description": "Modify specie by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The specie id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Specie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Specie"
        ],
        "description": "Delete specie by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The specie id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
...
  }
}